media: soc_camera: rcar_vin: Fix alignment of clipping size
Since the Start Line Pre-Clip register, the Start Pixel Pre-Clip
register and the End Line Post-Clip register do not have restriction of
H/W to a setting value, the processing of alignment is unnecessary.
This patch corrects so that processing of alignment is not performed.

However, the End Pixel Post-Clip register has restriction of H/W which
must be an even number value at the time of the output of YCbCr-422
format. By this patch, the processing of alignment to an even number
value is added on the above-mentioned conditions.

The variable set to a register is as follows.

	- Start Line Pre-Clip register cam->vin_top
	- Start Pixel Pre-Clip register	cam->vin_left
	- End Line Post-Clip register pix->height
	- End Pixel Post-Clip register pix->width

This commit is based on upstream commit 90d322c24f95
	"media: soc_camera: rcar_vin: Fix alignment of clipping"

Signed-off-by: Koji Matsuoka <koji.matsuoka.xm@renesas.com>
Signed-off-by: Nhan Nguyen <nhan.nguyen.yb@renesas.com>