Revert "drm: rcar-du: Fix H/V sync signal polarity configuration"
This reverts commit 0d2b7767611f ("drm: rcar-du: Fix H/V sync
signal polarity configuration").

We got issue du crash when doing IT, The failed cases are random and not
always. After comparing with cip3 which gave a good IT result, we found
that the commit 0d2b7767611f ("drm: rcar-du: Fix H/V sync
signal polarity configuration") made issue.

More investigation is needed, but we revert it for now.

Signed-off-by: Binh Nguyen <binh.nguyen.uw@renesas.com>
Signed-off-by: Duy Dang <duy.dang.yb@renesas.com>