sched/fair: Add group_misfit_task load-balance type
To maximize throughput in systems with asymmetric cpu capacities
(e.g. ARM big.LITTLE) load-balancing has to consider task and cpu
utilization as well as per-cpu compute capacity when load-balancing
in addition to the current average load based load-balancing policy.
Tasks with high utilization that are scheduled on a lower capacity cpu
need to be identified and migrated to a higher capacity cpu if possible
to maximize throughput.

To implement this additional policy an additional group_type
(load-balance scenario) is added: group_misfit_task. This represents
scenarios where a sched_group has one or more tasks that are not
suitable for its per-cpu capacity. group_misfit_task is only considered
if the system is not overloaded or imbalanced (group_imbalanced or
group_overloaded).

Identifying misfit tasks requires the rq lock to be held. To avoid
taking remote rq locks to examine source sched_groups for misfit
tasks, each cpu is responsible for tracking misfit tasks themselves and
update the rq->misfit_task flag. This means checking task utilization
when tasks are scheduled and on sched_tick.

Cherry-picked from https://patchwork.kernel.org/patch/10506537/

cc: Ingo Molnar <mingo@redhat.com>
cc: Peter Zijlstra <peterz@infradead.org>

Signed-off-by: Morten Rasmussen <morten.rasmussen@arm.com>
[Changes to use task_util() instead of task_util_est() on v4.14]
Signed-off-by: Gaku Inami <gaku.inami.xh@renesas.com>
Signed-off-by: Hien Huynh <hien.huynh.px@renesas.com>