clk: renesas: rcar-gen3: Fix SD divider setting
This patch is based on commit:
<2de4f963524d2f6f0c0373f684fb7f0af997d4a7>
in repository kernel/git/horms/renesas-bsp.git to fix the SD divider
settiing for corresponding to the change in the HS200/HS400 mode.

Note in the referrence commit:
- In H3/M3, the clock in HS400 mode refers to SDnH clock
  (others are SDn clock). Therefore, only the division value of SD is set
  as struct sd_div_table.div of cpg_sd_div_table[].
- In M3N, HS400 mode and HS200 mode use the same clock setting.

Therefore, apply this patch to support RZ/G2M, G2N and G2H(in future).

Signed-off-by: Takeshi Kihara <takeshi.kihara.df@renesas.com>
Signed-off-by: Hien Huynh <hien.huynh.px@renesas.com>