media: ov5645: change the way of setting virtual channel
This commit is created to change the way of setting virtual channel
for OV5645.
Currently, in previous patch, we use module parameter to set virtual
channel. Limitation of this way is can not use if system have more than
2 cameras OV5645.
To avoid this limitation, we will get data from devicetree with
"virtual-channel" property in OV5645 device node.
Example:
To use OV5645 in virtual channel 1, we define in devicetree
as below:
	ov5645 {
		....
		virtual-channel = /bits/ 8 <1>;
		port {
			...
		};
	};

Signed-off-by: Hien Huynh <hien.huynh.px@renesas.com>