gpio: rcar: Prevent to force the IRQ handler as threaded on RT system
When kernel is booting with CONFIG_IRQ_FORCED_THREADING flag, the
gpio_rcar_irq_handler() will be forced as threaded and then will
generate a WARNING relating to interrupt handlers. To prevent to
force the interrupt handler as threaded, we will add IRQF_NO_THREAD
flag when requesting an interrupt for gpio-rcar driver.

Signed-off-by: thanhbui <thanh.bui.fz@renesas.com>