spi: sh-msiof: Add slave mode support
Add slave mode support to the MSIOF driver for RZ/G1.
To configure msiof for spi slave mode, add property "spi-slave"
to devicetree nodes; otherwise, master will be used as default.

NOTE: Currently, MSIOF SPI slave driver only supports the transmission of
messages with a size that is known in advance (the hardware can provide
an interrupt when CS is deasserted before, though).

Ex: If the SPI master sends a shorter message, the slave won't receive it.
Or, when the SPI master sends a longer message, the slave will
receive the first part, and the rest will remain in the FIFO.

Based on commit cf9e4784f3bde3e4749163384f27450ddffe746c upstream.

Signed-off-by: Loc Vu <loc.vu.zn@renesas.com>