pinctrl: pinctrl-rzg2l: prevent to force the interrupt handler as thread
On RT system, when kernel is booting with CONFIG_IRQ_FORCED_THREADING
flag, the rzg2l_pinctrl_irq_handler() will be forced as threaded and
then will generate a WARNING MESSAGE related to interrupt handlers.
To prevent to force the interrupt handler as threaded, we will add
IRQF_NO_THREAD flag when requesting an interrupt for gpio-rcar driver.

Signed-off-by: Hien Huynh <hien.huynh.px@renesas.com>