drivers: spibsc: workaround mount MTD issue in 4bit data transfer
In mounting MTD process, the flash is still busy after some read
commands so the next read command returns wrong value, it causes
error when read value in MTD partition after mounting it.
Waiting a duration in each read for the flash's ready can be a
workaround for this issue.

Signed-off-by: Quynh Nguyen <quynh.nguyen.xb@renesas.com>