clocksource: mtu3: Add support for complementary pwm
In complementary pwm mode, MTU3 can export 2 different pwm signals
for pins of mtu3 channel x in pairs (MTIOCxA and MTIOCxC) or
(MTIOCxB and MTIOCxD) in opposite levels corresponding to output = 0
and output = 1.
Deadtime, the time which 2 pwm signals simultaneously have a same level
in one cycle in nanoseconds which can be configured in devicetree
as the following format.

	pwm_complementary = <channel output deadtime_ns>;

Ex: To configure channel 3 output complementary signals in MTIOC3B
and MTIOC3D with deadtime is 1000000 nanoseconds.

	pwm_complementary = <3 1 100000>;

To suppress dead time between 2 outputs, change deadtime to 0
in devicetree.

Signed-off-by: Loc Vu <loc.vu.zn@renesas.com>