arm64: dts: renesas: Add HiHope RZ/G2M board v3.0 with idk-1110wr display
The HiHope RZ/G2M v3.0 is advertised as compatible with panel idk-1110wr
from Advantech, however the panel isn't sold alongside the board.
A new dts, adding everything that's required to get the panel to
work the HiHope RZ/G2M v3.0, is the most convenient way to support the
HiHope RZ/G2M v3.0 when it's connected to the idk-1110wr.

Signed-off-by: Khoi Nguyen <khoi.nguyen.aj@renesas.com>
Signed-off-by: hienhuynh <hien.huynh.px@renesas.com>