net: can: rcar: rcar_canfd: Corrected the interrupt registration procedure of Channel interrupt.
Currently, in RZG2L platform, the Channel interrupt for each channel only correct
if we enable both 2 channels, when we disabled channel-0 and enabled channel-1 only,
the transmission data was failed.

It occurred because the Channel interrupt for channel-1 only is not correct.
This patch to correct the interrupt registration procedure of Channel interrupt
and resolved this issue.

Signed-off-by: Hiep Pham <hiep.pham.zy@renesas.com>