rcar-du: rcar_rgb_dummy: add RGB Dummy Encoder driver support
Currently, in DU driver, DPAD output must be connected to a bridge.
If it connects directly to a panel, DU driver will not be registered.

So this commit is created to implement connection between DPAD and
Panel by adding a dummy RGB Encoder driver.
This driver will become a bridge connection between DPAD output and
RGB Panel.

Signed-off-by: hienhuynh <hien.huynh.px@renesas.com>