drm: rcar-du: Add setting to PnALPHAR register on Gen3
Refer from commit id
<fcb34fe338cbde0a64919430733541035f20a784>
in renesas-rcar/linux-bsp repository with branch v4.14.75-ltsi/rcar-3.9.11
to apply it for RZ/G2 SoCs.

In Gen3, when Alpha blend is enabled in the PnMR register,
depending on the initial value of the PnALPHAR register,
either channel of DU might be black screen.
Therefore, this patch prevents the black screen by setting
the PnALPHAR register to all 0.

Signed-off-by: LUU HOAI <hoai.luu.ub@renesas.com>
Signed-off-by: hienhuynh <hien.huynh.px@renesas.com>