clocksource: renesas-mtu3: Improve register access method
In the old way, we cannot access all MTU3 registers properly.
Thus, we separated read/write register functions of channels to 8-bit,
16-bit and 32-bit.

Developer must know type of the registers they want to interact with
to choose suitable functions. Every functions can return -EINVAL to
announce that the input channels or offsets are wrong. Registers of
channel 5 (MTU5) are also supported in this commit, but be aware that
their structure and macros are totally different from other channels.

With shared registers, we still keep the old manner and add more
macros of their offsets.

Signed-off-by: Loc Vu <loc.vu.zn@renesas.com>