clocksource: renesas-mtu3: Make setting cnt inputs easy to understand
In the old way, we write 0 or 1 to "in_count_X_inputs" attributes to switch
between pairs of input clks, but user could be confused when using
this manner.

To make it more clear, we write directly name of the pair of input pins
which are "MTCLKA-MTCLKB" or "MTCLKC-MTCLKD" to change count sources.

For example: To change current pins "MTCLKA-MTCLKB" to "MTCLKC-MTCLKD"
of 32-bit counter, simply type:

  echo "MTCLKC-MTCLKD" > in_count_32bit_inputs.

Signed-off-by: Loc Vu <loc.vu.zn@renesas.com>