pinctrl: pinctrl-rzg2l: change irq_disable to irq_shutdown function
This commit is created to change irq_disable to irq_shutdown function.
Currently, we used a pair of irq_unmask/irq_mask to control GPIO IRQ.
So that using irq_disable is not good. It should be changed to
irq_shutdown.
irq_shutdown will be called when any drivers free its GPIO IRQ.

Signed-off-by: hienhuynh <hien.huynh.px@renesas.com>