soc: renesas: Fix mistake that overwrite revision of wrong target
The product revision information is overwritten in case current SoC
is RZG2UL. However, the strcmp() function was used wrongly when trying
to detect RZG2UL. Due to this mistake, product revision of other SoCs
are overwritten instead.
strcmp returns 0 when the strings match, so its result must be reverted.

Signed-off-by: Hung Tran <hung.tran.jy@renesas.com>
Signed-off-by: Hien Huynh <hien.huynh.px@renesas.com>