arm64: dts: renesas: rzg2ul-smarc: add RST and EN pins for OV5645 camera sensor
This patch is created to add RST and EN for OV5645 camera sensor in
RZ/G2UL smarc platform.
Enable pin of OV5645 is muxed with eth0/MDIO pin and chosen by SW1-3.
- If SW1-3 is 0, eth0/MDIO pin will be selected and enable pin of
OV5645 is always pulled up. So that choose a dummy pin to ensure probing
successfully for OV5645. Use P8_1 here because it does not connect to
any pin.
- If SW1-3 is 1, P4_4 will be selected as enable pin of OV5645.

Signed-off-by: hienhuynh <hien.huynh.px@renesas.com>