arm64: dts: renesas: r9a08g045: Correct interrupts setting of MTU3
This patch corrects the type of MTU3 interrupts from level
to edge and changes the interrupt-names for MTU3 driver
can read and enable them properly.

Some interrupts of MTU3 and GPT are mixed on RZ/G3S, so we
add a macro to define when switching to enable MTU3's interrupts.
By default, GPT's interrupts are priorized.

Signed-off-by: Loc Vu <loc.vu.zn@renesas.com>