dts: renesas: Add macro to enable/disable all interrupts of MTU3
When IRQs are used a long with PWM/Timer functions, the number of IRQ events
could be huge during run time because they are generated after compare-match
events.
It's better that we suppose example of setting all MTU3 interrupts,
but disable by defaults. Users will follow it and do their customization.

Signed-off-by: Loc Vu <loc.vu.zn@renesas.com>