i3c: renesas-i3c-master: Set data transfer for transmitting data via Command Descriptor
For transmitting data less than 4 bytes, the Immediate Transfer Command
will be used, and the data needs to be written to the most significant
DWORD of the Command Descriptor.

For transmitting data greater than 4 bytes, the Regular Transfer Command
will be used, and the data needs to be written to the buffer data.

Signed-off-by: Au Dang <au.dang.jx@renesas.com>