i3c: renesas-i3c-master: Correct Standard Bit Rate setting
Currently, the high/low period values are determined solely by bit rate
regardless of DSBRPO bit setting. This may cause situation that one of
the high/low period is doubled whilst the other one is not.

This patch is to fix the issue by making high/low period calculation
base on DSBRPO setting.

Signed-off-by: Quynh Nguyen <quynh.nguyen.xb@renesas.com>