arm64: dts: renesas: rzg3s-dev-common: Enable GPT0/POEG
Adds example of usage for MTU3/POE3 and GPT/POEG for
verification. For MTU3, MTIOC7A/C can be used as complementary PWM
outputs, and POE3 is used to control enabling of these pins.

For GPT, GPT0 is selected to output signals via its pins, and POEG{C,D}
are used to control enabling of these pins.

Signed-off-by: Loc Vu <loc.vu.zn@renesas.com>
Signed-off-by: Quynh Nguyen <quynh.nguyen.xb@renesas.com>
Signed-off-by: hienhuynh <hien.huynh.px@renesas.com>