i3c: renesas-i3c-master: Workaround can't read/write I3C after S2R
I3C device needs to be Dynamic Address Assignmet (DAA) again after S2R
before being able to use, but the current I3C framework does not support
any APIs to do it, so DAA should be done once in resume function to
workaround this issue.
The current resume function works in NOIRQ phase so DAA cannot be
performed in it since the IRQ callback function cannot be invoked in
NOIRQ phase. Therefore, DAA can only be done in the later phase of S2R.

Signed-off-by: Quynh Nguyen <quynh.nguyen.xb@renesas.com>