dmaengine: sh: rz-dmac: Allow for multiple DMACs
dma_request_channel() calls into __dma_request_channel() with
NULL as value for np, which won't allow for the selection of the
correct DMAC when multiple DMACs are available.

Switch to using __dma_request_channel() directly so that we can
choose the desired DMA for the channel. This is in preparation
of adding DMAC support for the Renesas RZ/V2H(P) and similar SoCs.

Refer from commit id:
https://patchwork.kernel.org/project/linux-renesas-soc/patch/20250218234305.700317-6-fabrizio.castro.jz@renesas.com/

Signed-off-by: Fabrizio Castro <fabrizio.castro.jz@renesas.com>
Signed-off-by: Tranh Ha <tranh.ha.xb@renesas.com>