media: rzg2l-cru: add frame skipping v4l2 control support for CRU
Currently, there are some camera sensors need time to be stable when
starting streaming such as OV5645,...
So this commit is created to add a user-class private V4L2 ctrl
named "frame skipping" to enable/disable skipping 1st several frames
when begin starting capture to let camera device go to stable state.
We set number of skipping frames to 3.
Default of this ctrl is disabled.

Signed-off-by: hienhuynh <hien.huynh.px@renesas.com>
Signed-off-by: Nghia Vo <nghia.vo.zn@renesas.com>