driver: thermal: rzt2h_thermal: Disable TSU driver access to OTP
In the RZ/T2H, the Linux temperature sensor driver
directly accesses the OTP. However, because the OTP
stores security information, it must be placed
in the secure world.

If the OTP is placed in the secure world,
Linux in the normal world will not be able to access it.
Therefore, the temperature sensor driver must use SMC
to obtain information from the OTP.

Signed-off-by: Huy Nguyen <huy.nguyen.wh@renesas.com>