drivers: pwm: gpt-rzg2l: Fix wrongly generated output pulse number.
- In the last pulse, the overflow irq has raised and set the CST
bit to 0 in the GTCR register but the GTCNT Register is still
counting and has a value before stopping. Therefore, it is
necessary to use the OADFLT/OBDFLT bit in GTIOR to control the
output value when stopping. Also use compare irq to stop signal
at small duty cycle.
- Reset the counter value (GTCNT) at the end and when switching
modes to avoid the situation of incorrectly generated output
pulse number next time.

Signed-off-by: Tranh Ha <tranh.ha.xb@renesas.com>
Signed-off-by: ductrinh <duc.trinh.wz@renesas.com>