arm64: dts: renesas: r9a09g077m44-dev.dts: Enable TBS, multiqueues for GMAC1,2
To support taprio on HW, GMAC need to enable multiqueues TX.
The total fifo size of 3 queues is 8192 bytes, if enable 3 queues with
store-and-forward mode, it's not enough for large packets, which would
trigger fifo overflow frequently, so we need to set DMA to thresh mode to
enable all 3 queues.

To support etf qdisc hw offload, GMAC need to enable TBS.
When enable TBS, GMAC will use Enhanced descriptor for DMA, which will
increase size of DMA descriptor. If we use DMA thresh mode when
enable TBS, we must increase DMA threshold for compatible with
DMA Enhanced descriptor.

This patch enable TBS, modify DMA threshold and enable 3 TX queues for
GMAC1,2 to support taprio and ETF in HW.

Signed-off-by: Long Luu <long.luu.ur@renesas.com>
Signed-off-by: Hoa Huynh <hoa.huynh.ud@renesas.com>