sound: rz-ssi: workaround to set TX and RX at the same time
Based on HW manual, set 1 at the same time when using SSIF-2 as
transmission and reception.
If not, sometimes playback and capture will fail (no interrupt raised)
when running both continuously.
So, add a workaround to clear REN and TEN bits then set 1 to REN and TEN
bits when full duplex is used

Signed-off-by: Hien Huynh <hien.huynh.px@renesas.com>