driver: i2c: busses: i2c-riic: Avoid NULL pointer dereference in ISRs
Some case the num_slave variable is set to -1. However, only
slave[0:2] are assigned to NULL which make the condition if
(slave[num_slave]) become unstable and lead to NULL pointer dereference.
So, a condition to check num_slave > -1 is added to avoid accessing
slave[-1].

Update unreg_slave() to clear ICIER when board change from slave mode
to master mode and clear start interrupt to make the code cleaner

Signed-off-by: Son Hoang <son.hoang.aj@renesas.com>
Signed-off-by: Tu Nguyen <tu.nguyen.xg@renesas.com>