i2c: busses: i2c-rsci: Add suspend/resume support
Add suspend/resume support for the i2c-rsci driver. This is necessary for
the Renesas RZ/G3E SoC which support suspend to deep sleep state where power
to most of the SoC components is turned off. As a result the I2C controller
needs to be reconfigured after suspend/resume. For this, the reset line
was stored in the driver private data structure as well as i2c timings.
The reset line and I2C timings are necessary to re-initialize the
controller after resume.

Signed-off-by: Tranh Ha <tranh.ha.xb@renesas.com>
Signed-off-by: anhvanmai <anh.mai.yg@bp.renesas.com>