pwm: rzg2l-gpt: Accept requests for too high period length
The period setting is shared for each pair of PWM channels. So if the
twin channel is in use, the period must not be changed. According to the
usual practise to pick the next smaller possible period, accept a
request for a period that is bigger than the unchangable value.

Signed-off-by: Uwe Kleine-König <u.kleine-koenig@baylibre.com>
Signed-off-by: Uwe Kleine-König <ukleinek@kernel.org>
Signed-off-by: Y Pham <y.pham.yg@bp.renesas.com>