drm: rz-du: Support panels connected directly to the DPAD outputs
Refer commit "drm: rcar-du: Support panels connected directly to the DPAD outputs".

The RZ DU driver assumes that a bridge is always connected to the DU
output. This is valid for the LVDS and HDMI outputs, but the DPAD
outputs can be connected directly to a panel, in which case no bridge is
available.

To support this use case, detect whether the entities connected to the
DU DPAD outputs are encoders or panels based on the number of ports of
their DT node, and retrieve the corresponding type of DRM objects. For
panels, additionally create panel bridge instances.

Signed-off-by: hienhuynh <hien.huynh.px@renesas.com>
Signed-off-by: Nhan Nguyen <nhan.nguyen.uf@bp.renesas.com>