pci: controller: pcie-rzt2: Countermeasures the PCIe unit in the LSI becomes deadlocked
Add PCIe support for RZ/T2N based on the implementation for RZ/V2M/V2MA
("drivers: pci: controller: v2m/v2ma: Countermeasures the PCIe unit in
the LSI becomes deadlocked")

In the following Lane connection conditions, if Link Down/Hot Reset
occurs, the mechanismthat controls PCIe communication processing
(LTTSM control) stops. After that, it will not be possible to establish
communication with the opposing device.

This patch add to kernel driver to countermeasure that. When a reset
request is received from the OSS department, If detect communication
interrupt is detected, reset the Link and Phy on the appropriate Lane
(resource) side.

Signed-off-by: Tai Huynh <tai.huynh.cp@renesas.com>
Signed-off-by: Duan Le <duan.le.xa@renesas.com>