clk: renesas: r9a08g045: Add clocks and resets support for PCIe
Add clocks and resets for the PCIe IP available on the Renesas RZ/G3S SoC.
The clkl1pm clock is required for PCIe link power management (PM) control
and should be enabled based on the state of the CLKREQ# pin. Therefore,
mark it as a no_pm clock to allow the PCIe driver to manage it during link
PM transitions.

Tested-by: Wolfram Sang <wsa+renesas@sang-engineering.com>
Signed-off-by: Claudiu Beznea <claudiu.beznea.uj@bp.renesas.com>
Signed-off-by: Dat <dat.vo-thanh@banvien.com.vn>
Link:https://lwn.net/ml/all/20250704161410.3931884-3-claudiu.beznea.uj@bp.renesas.com/