media: rzg2l-cru: Set auto recovery mode when AXI Bus is congested
Sometimes, AXI bus congestion caused the image data being corrupted.
This patch is created to set auto recovery mode when AXI Bus is
congested.
It will stop AXI master and clear the FIFO.
After recovery, transfer resumes to the start address of the Memory Bank
to which data was being transferred to.

Because the detection of this issue is recognized by looping frame with
a same HW memory bank, we update minimum buffer to 2 buffers to detect
this issue.

Signed-off-by: hienhuynh <hien.huynh.px@renesas.com>
Signed-off-by: Nghia Vo <nghia.vo.zn@renesas.com>