drivers: watchdog: rzg2l_wdt: Configure CPG_WDTRST_SEL register
Currently, the watchdog driver relies on TF-A/U-Boot to configure the
CPG_WDTRST_SEL register. This register must be set correctly to
ensure a reset request is issued upon watchdog timer (WDT) underflow.

Now that the driver has access to the syscon handle for CPG, configure
CPG_WDTRST_SEL directly instead of depending on firmware. This
improves robustness by ensuring the required configuration is applied
within the driver itself.

Signed-off-by: Nguyen Phuc Dung <dung.nguyen.ks@bp.renesas.com>