rtc: renesas-rtca3: Increase timeout when clearing RCR2.START
On some RZ/G3S Smarc Boards, the DEFAULT_TIMEOUT (150 µs) is not
sufficient for the hardware to clear the START bit in RCR2 during
initial setup. This may cause RTC failed to probe on the first boot
power-on.

This patch applies a specific timeout which practically works for
different RZ/G3S Smarc boards.

Signed-off-by: Hien Dang <hien.dang.xf@renesas.com>