mtd: rawnand: Add NAND controller driver for RZ/T2N
Add support for the NAND Flash controller found on Renesas RZ/T2N
SoCs.

This driver is independent from the existing rzn1-nand driver, as
the register set and operation flow are different.

The overall structure is based on the Cadence NAND controller driver
which is adapted to the RZ/T2N register set.

Signed-off-by: Anh Ly <anh.ly.yb@renesas.com>