tty: serial: rz-sci: Use RDR FFER/FPER to detect error flags for FIFO mode
Based on HWUM, the FER, PER bits of the CSR register indicate general error states when using FIFO.
If the data in the FIFO has one or a few bytes with frame or parity errors, FER/PER will be set.
So if using FIFO then check the FFER/FPER flag of RDR register instead of CSR register.

Signed-off-by: anhvanmai <anh.mai.yg@bp.renesas.com>