gpu: drm: rz-du: use optional and share for reset control
Using the devm_reset_control_get_optional_shared for reset control
to ensure that get the reset signal is optional for RZ/T2H series and
the reset signal is shared for RZ/G2L series.

This is added to fixed this commit: cf14d706

Update to share reset resource for G2L/LC/UL devices.

Signed-off-by: ThanhLe <thanh.le-tan@banvien.com.vn>
Signed-off-by: Nhan Nguyen <nhan.nguyen.uf@bp.renesas.com>