arm64: dts: renesas: rzg3s-smarc: Resolve gpt0 pinmux conflict with CAN PMODs
The pins defined for the 'gpt0_pins' node (GTIOC0A - P13_0 and
GTIOC0B - P13_1) share physical pins with the CAN interface
when routed to PMOD1 or PMOD2.

Previously, the 'gpt0_pins' node was defined unconditionally,
which resulted in a pinmux conflict error if either
'SW_GPIO_CAN_PMOD1' or 'SW_GPIO_CAN_PMOD2' was enabled (SW_ON).

This patch wraps the 'gpt0_pins' definition in a C preprocessor
directive. The pins will now only be allocated to the GPT timer
if both CAN PMOD switches are in the 'SW_OFF' state, thus
resolving the conflict.

Signed-off-by: Nguyen Phuc Dung <dung.nguyen.ks@bp.renesas.com>