irqchip: irq-renesas-rzv2h: Use IRQF_NO_THREAD for GPT irqs in RT kernel
When a Dead Time error occurs and the POEG is enabled, an IRQ is triggered
to disable the outputs on both GPT0 channels A and B. In RT environment,
the IRQ handlers are threaded by default and may cause delay which result
in unexpected noisy waveforms.

Fix this by using IRQF_NO_THREAD for GPT irqs request.

Signed-off-by: Sang Nguyen <sang.nguyen-hoang@banvien.com.vn>