tty: serial: rz-sci: Only clear the TDRE flag when using FIFO mode
Based on Figure 7.3-15 Example of Serial Transmission Flowchart in
Asynchronous Mode (Non-FIFO Selected) in the User’s Manual: Hardware
(Additional document). Data transmission in non-FIFO mode does not
include clearing the TDRE flag. Clearing the TDRE flag must be done
if FIFO mode is selected (Figure 7.3-17).

According to the TDRE bit description, it will be automatically
cleared when data is written to the TDR register.

Signed-off-by: anhvanmai <anh.mai.yg@bp.renesas.com>