pwm: rzg2l-gpt: Update normal_output and single/double buffer output
In single/double buffer mode, the first output period is
incorrect when the mode is enabled.
Likewise, updating the duty cycle while the PWM is running
causes the first output period to be incorrect.

This patch updates the duty cycle and single/double buffer
settings in the overflow interrupt handler.

Signed-off-by: Sang Nguyen <sang.nguyen-hoang@banvien.com.vn>
Signed-off-by: Y Pham <y.pham.yg@bp.renesas.com>